requirements.txt
streamlit==1.49.1
streamlit-webrtc==0.63.4
python-dotenv==1.1.1
pymongo==4.14.1
passlib==1.7.4
bcrypt==4.0.1
requests==2.32.5
murf==2.0.2
elevenlabs==2.13.0
soundfile==0.13.1
numpy==2.2.6
av==14.4.0
elevenlabs